reference/data_crawling/requirements.txt
# Web Crawlers Dependencies

# HTTP requests
requests>=2.28.0

# HTML parsing
beautifulsoup4>=4.12.0

# Logging
loguru>=0.7.0

# Optional: for type hints
typing-extensions>=4.0.0

# Email handling
imap-tools>=1.0.0
python-dotenv>=1.0.0

# GCP sdk
google-cloud-storage

# Retry mechanism
tenacity